+++ app/js/preloads/follow.js
@@ -30,10 +30,14 @@ const clickableInputs = [
     "label[for]:not([for=\"\"])",
     "input[type=\"submit\"]",
     "input[type=\"color\"]",
+    "input[type=\"file\"]",
+    "input[type=\"image\"]",
+    "input[type=\"reset\"]",
     "summary"
 ]
 const textlikeInputs = ["input:not([type=\"radio\"]):not([type=\"checkbox\"])"
-    + ":not([type=\"submit\"]):not([type=\"button\"]):not([type=\"color\"])",
+    + ":not([type=\"submit\"]):not([type=\"button\"]):not([type=\"color\"])"
+    + ":not([type=\"file\"]):not([type=\"image\"]):not([type=\"reset\"])",
 "textarea",
 "select"]